Move FluentParserStream to fluent/syntax/stream.py
